Add TextView.textColor ext
Former-commit-id: 97ab666326920f32485df03bacdc8758aef7b8b0 [formerly 4dac794d6f82cca71af821eed75485f41f6811aa]
Former-commit-id: 4fff734ceaf6d1f732f2972637707983ca8c0fa8

+++ id/src/main/java/com/simprints/id/activities/settings/fragments/moduleselection/tools/ModuleChipHelper.kt
@@ -44,7 +44,7 @@ class ModuleChipHelper(private val context: Context, private val listener: ChipC
     private fun createChipDrawable(): ChipDrawable {
         return ChipDrawable.createFromResource(context, R.xml.module_selection_chip).apply {
             setTextAppearanceResource(R.style.SimprintsStyle_TextView_Chip)
-            val cornerSize = context.dimen(R.dimen.chip_corner_size_module_selection)
+            val cornerSize = context.dimen(R.dimen.chip_corner_size_module_selection).toFloat()
             shapeAppearanceModel =
                 ShapeAppearanceModel.builder().setAllCorners(CornerFamily.CUT, cornerSize).build()
         }
